requirements/users.txt
# Users service specific requirements
-r base.txt

# django-ninja and JWT for users service
django-ninja>=1.0.0
django-ninja-jwt>=5.3.0

# TODO: Add users-specific dependencies from users/pyproject.toml
# Review the actual pyproject.toml and migrate dependencies here
